+++ _demo/AngularMicrophone/android/app/src/main/java/com/mozartec/capacitor/microphone/angular/MainActivity.java
@@ -1,5 +1,0 @@
-package com.mozartec.capacitor.microphone.angular;
-
-import com.getcapacitor.BridgeActivity;
-
-public class MainActivity extends BridgeActivity {}