+++ _demo/vue/capacitor.config.ts
@@ -1,0 +1,9 @@
+import type { CapacitorConfig } from '@capacitor/cli';
+
+const config: CapacitorConfig = {
+  appId: 'tech.idsynergy.microphonedemo',
+  appName: 'Microphone Demo',
+  webDir: 'dist',
+};
+
+export default config;